frame_service/requirements.txt
grpcio
grpcio-tools
opencv-python # 4.9.0.80
pymongo
aiokafka
Pillow
apscheduler
minio